+++ CHANGES.md
@@ -4,6 +4,7 @@
 * refactored Settings to make smaller methods
 * simplfied `composer.json`, removed autoload section and no need to require `composer/installer`
 * more PHP linting fixes
+* fixed `class Rest_Update` for PHP 5.3 compatibility, thanks @epicfaace
 
 #### 6.3.5 / 2017-06-29
 * hotfix to `composer.json` to remove classmap and files, I think I messed something up.